// src/app/models/pdt.ts
export interface PDT {
    id: string;
    createdDate: string;
    createdTimestamp: string;
    empName: string;
    shortTermGoals: string;
    mediumTermGoals: string;
    longTermGoals: string;
    developmentNeeds: string;
    actionPlan: string;
    empSignature: string;
    superSignature: string;
}